+++ NEWS.rst
@@ -5,6 +5,13 @@ Unreleased
 
 Breaking Changes
 ------------------------------
+
+* `defmacro` and `require` can now define macros locally instead of
+  only module-wide.
+
+  * `hy.eval`, `hy.macroexpand`, `doc`, and `delmacro` don't work with
+    local macros (yet).
+
 * When a macro is `require`\d from another module, that module is no
   longer implicitly included when checking for further macros in
   the expansion.